breakpad: Fix an assertion encountered in UntypedMDRVA::Copy().
Fix an assertion where a zero-length buffer was being passed to
UntypedMDRVA::Copy().  This will occur if WriteFile() is given a file whose
size is a multiple of the temporary buffer size, and is presumably what is
causing google-breakpad unit tests to fail only on stumpy.

This is the same fix as was needed for core2md (see
"http://gerrit.chromium.org/gerrit/#change,3358").

BUG=chromium-os:21402, chromium-os:17032
TEST=Ran unittests.

Change-Id: I86cb395805aa86e22b65badc399b9be2784092f1
Reviewed-on: http://gerrit.chromium.org/gerrit/9920
Tested-by: Michael Krebs <mkrebs@chromium.org>
Commit-Ready: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Ken Mixter <kmixter@chromium.org>