google-breakpad: Fix test addresses to use uintptr_t instead of u_int64_t.
When a variable is used to set (and lookup) MappingInfo's "start_addr"
field, it needs to match types -- which is "uintptr_t".  Updating the 'make'
that's used for building apparently caused the 32-bit "char *" to
sign-extend when cast up to a u_int64_t -- maybe because pointers were
unsigned before and then changed to be signed?

BUG=chromium-os:25355
TEST=Ran Breakpad unittests

Change-Id: I79ddd447e41e44a6d33ceec84a57671636132cdc
Reviewed-on: https://gerrit.chromium.org/gerrit/14859
Tested-by: Michael Krebs <mkrebs@chromium.org>
Commit-Ready: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>