google-breakpad: Fix for putting main module as first one in minidump
The first module in a minidump is expected to be for the main executable.
We used to assume that /proc/<pid>/maps always showed that one first, but in
some cases that is no longer true (see comment #7 of the bug).  So this
change makes use of the entry point stored in auxv to make sure we put the
correct module first.

BUG=chromium-os:25355
TEST=Ran Breakpad unittests for host and target
CQ-DEPEND=I763ce0196852775283c826e19d0ec86cb0516769

Change-Id: Iaa814ff64247953ba0dfe89e09bd7ab946d8e84c
Reviewed-on: https://gerrit.chromium.org/gerrit/18917
Tested-by: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Michael Krebs <mkrebs@chromium.org>